fix: remove crashReporterRenderer.sendSync() workaround for init()
